get_stream/requirements.txt
requests
Mastodon.py